Big Makefile redo with cleaner common files and warnings cleanup

+++ apps/utils/Web/FileRepository.hpp
@@ -61,7 +61,7 @@ namespace web {
 
             fclose(fd);
 
-            return {buffer, size};
+            return {buffer, static_cast<size_t>(size)};
         }
 
     private:
